YUK-17915 IVVC Feeder deletion leads to UI errors

+++ yukon-client/yukon-web/src/com/cannontech/web/capcontrol/FeederController.java
@@ -234,22 +234,25 @@ public class FeederController {
     @CheckRoleProperty(YukonRoleProperty.CBC_DATABASE_EDIT)
     public String delete(@PathVariable int id, FlashScope flash) {
         CapControlFeeder feeder = feederService.get(id);
-        
-        try {
-            Integer parentId = feederDao.getParentSubBusID(id);
-            feederService.delete(id);
-            flash.setConfirm(new YukonMessageSourceResolvable(feederKey + ".delete.success", feeder.getName()));
-
-            if (parentId != null && parentId > 0) {
-                return "redirect:/capcontrol/buses/" + parentId;
+        Integer parentId = feederDao.getParentSubBusID(id);
+        boolean isCapbankAssigned = feederService.isCapBanksAssignedToZone(id);
+        if (isCapbankAssigned) {
+            flash.setError(new YukonMessageSourceResolvable(feederKey + ".delete.error", feeder.getName()));
+            return "redirect:/capcontrol/feeders/" + id;
+        } else {
+            try {
+                feederService.delete(id);
+                flash.setConfirm(new YukonMessageSourceResolvable(feederKey + ".delete.success", feeder.getName()));
+                if (parentId != null && parentId > 0) {
+                    return "redirect:/capcontrol/buses/" + parentId;
+                }
+            } catch (EmptyResultDataAccessException e) {
+                feederService.delete(id);
+                flash.setConfirm(new YukonMessageSourceResolvable(feederKey + ".delete.success", feeder.getName()));
+                // do nothing and return to orphan page
             }
-        } catch (EmptyResultDataAccessException e) {
-            feederService.delete(id);
-            flash.setConfirm(new YukonMessageSourceResolvable(feederKey + ".delete.success", feeder.getName()));
-            // do nothing and return to orphan page
+            return "redirect:/capcontrol/search/searchResults?cbc_lastSearch=__cti_oFeeders__";
         }
-        
-        return "redirect:/capcontrol/search/searchResults?cbc_lastSearch=__cti_oFeeders__";
     }
 
     @RequestMapping("feeders/{feederId}/capbanks/edit")
